[Cpp API Compatibility] Add some headeronly APIs (#78662)
