+++ paddle/phi/api/include/compat/c10/core/DeviceType.h
@@ -14,29 +14,15 @@
 
 #pragma once
 
-#include <cstdint>
-#include <functional>
+// If you modified DeviceType in this file, please also sync your changes into
+// torch/headeronly/core/DeviceType.h.
+#include <torch/headeronly/core/DeviceType.h>
+
 #include <ostream>
 
 #include "paddle/phi/common/place.h"
 
 namespace c10 {
-
-enum class DeviceType : int8_t {
-  CPU = 0,
-  CUDA = 1,
-  XPU = 12,
-  IPU = 18,
-  CUSTOM = 20,
-  PrivateUse1 = CUSTOM,
-};
-
-constexpr DeviceType kCUDA = DeviceType::CUDA;
-constexpr DeviceType kCPU = DeviceType::CPU;
-constexpr DeviceType kCUSTOM = DeviceType::CUSTOM;
-constexpr DeviceType kXPU = DeviceType::XPU;
-constexpr DeviceType kIPU = DeviceType::IPU;
-constexpr DeviceType kPrivateUse1 = DeviceType::PrivateUse1;
 
 inline phi::AllocationType DeviceTypeToPhi(DeviceType d) {
   switch (d) {
@@ -106,32 +92,3 @@ inline std::ostream& operator<<(std::ostream& os, DeviceType d) {
 }
 
 }  // namespace c10
-
-namespace std {
-template <>
-struct hash<c10::DeviceType> {
-  std::size_t operator()(c10::DeviceType k) const noexcept {
-    return std::hash<int>()(static_cast<int>(k));
-  }
-};
-}  // namespace std
-
-namespace at {
-using c10::DeviceType;
-using c10::kCPU;
-using c10::kCUDA;
-using c10::kCUSTOM;
-using c10::kIPU;
-using c10::kPrivateUse1;
-using c10::kXPU;
-}  // namespace at
-
-namespace torch {
-using c10::DeviceType;
-using c10::kCPU;
-using c10::kCUDA;
-using c10::kCUSTOM;
-using c10::kIPU;
-using c10::kPrivateUse1;
-using c10::kXPU;
-}  // namespace torch
